+++ lib/fireflutter.dart
@@ -33,12 +33,14 @@ export 'src/chat/widgets/chat.new_message.dart';
 export 'src/chat/widgets/chat.room_avatar.dart';
 export 'src/chat/widgets/chat.room_body.dart';
 export 'src/chat/widgets/chat.room_list_tile.dart';
+export 'src/chat/widgets/chat_room.list_view.dart';
 export 'src/chat/widgets/chat.total_no_of_new_message.dart';
+export 'src/chat/widgets/default_chat_room.blocked_user_list.screen.dart';
 export 'src/chat/widgets/default_chat_room.edit.dialog.dart';
 export 'src/chat/widgets/default_chat_room.invite.screen.dart';
 export 'src/chat/widgets/default_chat_room.list_view.dart';
 export 'src/chat/widgets/default_chat_room.member.dialog.dart';
-export 'src/chat/widgets/default_chat_room.members.screen.dart';
+export 'src/chat/widgets/default_chat_room.user_list.screen.dart';
 export 'src/chat/widgets/default_chat_room.screen.dart';
 
 export 'src/meetup/meetup.model.dart';