+++ lib/src/chat/widgets/chat.bubble.dart
@@ -14,16 +14,18 @@ import 'package:flutter/material.dart';
 class ChatBubble extends StatelessWidget {
   const ChatBubble({
     super.key,
+    required this.room,
     required this.message,
     this.onChange,
   });
 
+  final ChatRoom room;
   final ChatMessage message;
   final Function? onChange;
 
   @override
   Widget build(BuildContext context) {
-    return Container(
+    final bubble = Container(
       padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 8),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.start,
@@ -41,19 +43,15 @@ class ChatBubble extends StatelessWidget {
               cacheId: message.uid,
               size: 30,
               radius: 12,
-              onTap: () => UserService.instance
-                  .showPublicProfileScreen(
-                    context: context,
-                    uid: message.uid!,
-                  )
-                  .then(
-                    (value) => onChange?.call(),
-                  ),
+              onTap: () => mayShowPublicProfileScreen(context, message.uid!),
             ),
 
           const SizedBox(width: 8),
           // Chat message text. size 60%
-          if (message.text != null)
+
+          /// 여기서부터...
+          if (message.deleted ||
+              (message.text != null && message.text!.isNotEmpty))
             Container(
               constraints: BoxConstraints(
                   maxWidth: MediaQuery.of(context).size.width * 0.6),
@@ -68,11 +66,22 @@ class ChatBubble extends StatelessWidget {
                           : Colors.grey.shade200,
                       borderRadius: borderRadius(),
                     ),
-                    child: LinkifyText(
-                      message.text!
-                          .orBlocked(message.uid!, T.blockedChatMessage),
-                      style: const TextStyle(color: Colors.black),
-                    ),
+                    child: message.deleted
+                        ? Text(T.chatMessageDeleted.tr)
+                        : LinkifyText(
+                            message.text!
+                                .orBlocked(message.uid!, T.blockedChatMessage),
+                            style: const TextStyle(color: Colors.black),
+                          ),
+
+                    /// OLD code. commented out by @thruthesky at 2024-05-27
+                    // message.text != null && message.deleted == false
+                    //     ? LinkifyText(
+                    //         message.text!
+                    //             .orBlocked(message.uid!, T.blockedChatMessage),
+                    //         style: const TextStyle(color: Colors.black),
+                    //       )
+                    //     : Text(T.chatMessageDeleted.tr),
                   ),
                   if (message.hasUrlPreview) ...[
                     const SizedBox(height: 8),
@@ -100,12 +109,7 @@ class ChatBubble extends StatelessWidget {
               sync: true,
               size: 30,
               radius: 12,
-              onTap: () => UserService.instance
-                  .showPublicProfileScreen(
-                    context: context,
-                    uid: myUid!,
-                  )
-                  .then((value) => onChange?.call()),
+              onTap: () => mayShowPublicProfileScreen(context, myUid!),
             ),
 
           if (!message.mine) ...[
@@ -115,6 +119,85 @@ class ChatBubble extends StatelessWidget {
         ],
       ),
     );
+
+    /// 관리자가 아니고, 방장이 아니고, 나의 메시지가 아니면, 그냥 chat bubble 만 리턴
+    if (isAdmin == false && room.isMaster == false && message.mine == false) {
+      return bubble;
+    }
+
+    /// 나의 채팅 메시지이면, 삭제 등을 표시하기 위한 메시지 리턴
+    if (message.mine) {
+      return PopupMenuButton(
+        position: PopupMenuPosition.under,
+        onOpened: () => FocusScope.of(context).unfocus(),
+        offset: Offset(
+          message.mine ? MediaQuery.of(context).size.width : 0,
+          0,
+        ),
+        itemBuilder: (context) => <PopupMenuEntry<String>>[
+          PopupMenuItem(
+            value: Code.delete,
+            child: Text(T.chatMessageDelete.tr),
+          ),
+        ],
+        onSelected: (v) async {
+          switch (v) {
+            case Code.delete:
+              await message.delete();
+              break;
+            default:
+              break;
+          }
+        },
+        child: IgnorePointer(child: bubble),
+      );
+    }
+
+    /// 관리자이거나 방장이면, 팝업 메뉴 버튼을 리턴
+    return PopupMenuButton(
+      position: PopupMenuPosition.under,
+      onOpened: () => FocusScope.of(context).unfocus(),
+      offset: Offset(
+        message.mine ? MediaQuery.of(context).size.width : 0,
+        0,
+      ),
+      itemBuilder: (context) => <PopupMenuEntry<String>>[
+        PopupMenuItem(
+          value: Code.delete,
+          child: Text(T.chatMessageDelete.tr),
+        ),
+        PopupMenuItem(
+          value: Code.block,
+          child: Text(T.block.tr),
+        ),
+      ],
+      onSelected: (v) async {
+        switch (v) {
+          case Code.delete:
+            await message.delete();
+            break;
+          case Code.block:
+            await room.block(message.uid!);
+            break;
+          default:
+            break;
+        }
+      },
+      child: IgnorePointer(child: bubble),
+    );
+  }
+
+  mayShowPublicProfileScreen(BuildContext context, String uid) {
+    UserService.instance
+        .showPublicProfileScreen(
+          context: context,
+          uid: message.uid!,
+        )
+        .then(
+          // * when the profile is updated, the chat bubble does not reflect the change of profile
+          // * But for blocking/unblocking will be updated by this callback.
+          (value) => onChange?.call(),
+        );
   }
 
   borderRadius() {