+++ lib/src/user/widgets/user_avatar_update.dart
@@ -66,6 +66,7 @@ class _UserAvatarUpdateState extends State<UserAvatarUpdate> {
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () async {
+        ///
         final url = await StorageService.instance.uploadAt(
           context: context,
           path: "${User.node}/${user.uid}/${Field.photoUrl}",